+++ src/types/index.ts
@@ -15,6 +15,7 @@ export type Status =
   | { Connecting: Destination }
   | { Disconnecting: Destination }
   | { Connected: Destination }
+  | 'ServiceUnavailable'
   | 'Disconnected';
 
 export interface Destination {
@@ -24,29 +25,6 @@ export interface Destination {
 }
 
 export type Path = { Hops: number } | { IntermediatePath: string[] };
-
-// Type guards
-export function isConnected(
-  status: Status
-): status is { Connected: Destination } {
-  return typeof status === 'object' && 'Connected' in status;
-}
-
-export function isConnecting(
-  status: Status
-): status is { Connecting: Destination } {
-  return typeof status === 'object' && 'Connecting' in status;
-}
-
-export function isDisconnecting(
-  status: Status
-): status is { Disconnecting: Destination } {
-  return typeof status === 'object' && 'Disconnecting' in status;
-}
-
-export function isDisconnected(status: Status): status is 'Disconnected' {
-  return status === 'Disconnected';
-}
 
 // UI Types
 export interface ButtonProps
